Tentando resolver os problemas
