--- .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.jsch.core.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-org.eclipse.jsch.core.hasChangedDefaultWin32SshHome=true
