--- .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.jst.j2ee.webservice.ui.prefs
@@ -1,2 +1,0 @@
-areThereWebServices=false
-eclipse.preferences.version=1
