--- .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.mylyn.context.core.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-mylyn.attention.migrated=true
