--- .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.mylyn.monitor.ui.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-org.eclipse.mylyn.monitor.activity.tracking.enabled.checked=true
