--- .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.ui.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-showIntro=false
