--- .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.urischeme.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-processedSchemes=,eclipse+command,eclipse+mpc
