--- .metadata/.plugins/org.eclipse.jdt.core/javaLikeNames.txt
@@ -1,1 +1,0 @@
-java
