--- .metadata/.plugins/org.eclipse.jdt.core/savedIndexNames.txt
@@ -1,2 +1,0 @@
-INDEX VERSION 1.131+C:\Users\Usuario\Desktop\Orientacao\Projeto-OO-Lanna\.metadata\.plugins\org.eclipse.jdt.core
-1574843208.index
