--- .metadata/.plugins/org.eclipse.jdt.launching/.install.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<dirs>
-    <entry loc="C:\Program Files\Java\jdk-11.0.8" stamp="1600714658429"/>
-</dirs>
