--- .metadata/.plugins/org.eclipse.jdt.launching/libraryInfos.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<libraryInfos>
-    <libraryInfo home="C:\Program Files\Java\jdk-11.0.8" version="11.0.8"/>
-</libraryInfos>
