--- .metadata/.plugins/org.eclipse.tm.terminal.view.ui/.executables/data.properties
@@ -1,6 +1,0 @@
-#Mon Dec 07 00:05:45 BRT 2020
-0.Icon=C\:\\Program Files\\Git\\mingw64\\share\\git\\git-for-windows.ico
-0.Path=C\:\\Program Files\\Git\\bin\\sh.exe
-0.Translate=true
-0.Args=--login -i
-0.Name=Git Bash
