--- .metadata/.plugins/org.eclipse.ui.intro/introstate
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<state reopen="false"/>
