--- .metadata/version.ini
@@ -1,3 +1,0 @@
-#Mon Dec 07 00:05:42 BRT 2020
-org.eclipse.core.runtime=2
-org.eclipse.platform=4.17.0.v20200902-1800
