--- README.md
@@ -1,2 +1,0 @@
-# Projeto-OO-Lanna
- Projeto de OO para a turma FGA0158-T05
