--- Republica/src/aplicacao/Republica.java
@@ -1,22 +1,0 @@
-package aplicacao;
-
-import entidades.Despesas;
-import entidades.Pessoa;
-import entidades.Categoria;
-import entidades.Subcategoria;
-
-public class Republica {
-
-	public static void main(String[] args) {
-		
-	}
-	
-	public void relatorioProporcional() {
-		
-	}
-	
-	public void relatorioIgualitario() {
-		
-	}
-
-}
