--- Republica/src/restricoes/DadosPessoaisIncompletosException.java
@@ -1,13 +1,0 @@
-package restricoes;
-
-public class DadosPessoaisIncompletosException extends Exception {
-	
-	public DadosPessoaisIncompletosException() {
-		super(); 
-	}
-	
-	public DadosPessoaisIncompletosException(String msg) {
-		super(msg);
-	}
-
-}
