--- Republica/src/restricoes/RendimentoInvalidoException.java
@@ -1,13 +1,0 @@
-package restricoes;
-
-public class RendimentoInvalidoException extends Exception {
-	
-	public RendimentoInvalidoException() {
-		super(); 
-	}
-	
-	public RendimentoInvalidoException(String msg) {
-		super(msg);
-	}
-
-}
