+++ src/entidades/Subcategoria.java
@@ -1,0 +1,22 @@
+package entidades;
+
+public class Subcategoria {
+	
+	private String descricaoSubcategoria;
+
+	public Subcategoria(String descricaoSubcategoria) {
+		super();
+		this.descricaoSubcategoria = descricaoSubcategoria;
+	
+	}
+
+	public String getDescricaoSubcategoria() {
+		return descricaoSubcategoria;
+	}
+
+	public void setDescricaoSubcategoria(String descricaoSubcategoria) {
+		this.descricaoSubcategoria = descricaoSubcategoria;
+	}
+	
+	
+}
