+++ src/restricoes/DadosPessoaisIncompletosException.java
@@ -1,0 +1,13 @@
+package restricoes;
+
+public class DadosPessoaisIncompletosException extends Exception {
+	
+	public DadosPessoaisIncompletosException() {
+		super(); 
+	}
+	
+	public DadosPessoaisIncompletosException(String msg) {
+		super(msg);
+	}
+
+}
